# docs/agent/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Video Studio",
    "version": "1.0.3",
    "description": "Agent-facing API. Branch on the `code` field of an error, never on its prose.\nIn server mode every call carries `Authorization: Bearer avs_…`; scopes are read, produce, publish, admin.\nName the channel explicitly (body.channelId, ?channel= or X-AVS-Channel) — the active channel is only a default.\nSend an Idempotency-Key (or clientRef on a create) so a retry after a timeout cannot spend twice.\nPoll GET /events?after=<id> for progress; a WebSocket at /ws carries the same feed live."
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "code",
          "error"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine word. Branch on this, never on `error`."
          },
          "error": {
            "type": "string",
            "description": "Short reason, in the user's language unless ?lang= says otherwise."
          },
          "message": {
            "type": "string"
          },
          "hint": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Mint with `npm run token -- create` on the machine running the server."
      }
    }
  },
  "security": [
    {
      "bearer": []
    }
  ],
  "paths": {
...
    "/projects/{id}/publish": {
      "post": {
        "summary": "Upload the finished video. Defaults to private.",
        "tags": [
          "publish"
        ],
        "x-avs-scope": "publish",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "token_required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "scope_denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "default": {
            "description": "Known codes: not_found, video_not_ready, publish_not_connected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{id}/publishes": {
      "get": {
        "summary": "Where this video was published, and what happened.",
        "tags": [
          "publish"
        ],
        "x-avs-scope": "read",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "token_required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "scope_denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/publish/status": {
      "get": {
        "summary": "Which platforms are connected.",
        "tags": [
          "publish"
        ],
        "x-avs-scope": "read",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "token_required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "scope_denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}